+++ magda-int-test/src/test/scala/au/csiro/data61/magda/api/LanguageAnalyzerSpec.scala
@@ -111,10 +111,12 @@ class LanguageAnalyzerSpec extends BaseSearchApiSpec {
     testLanguageFieldSearch(termExtractor, test)
   }
 
+  def isAStopWord(term: String) = Generators.luceneStopWords.exists(stopWord => term.trim.equalsIgnoreCase(stopWord))
+
   def testLanguageFieldSearch(outerTermExtractor: DataSet => Seq[String], test: (DataSet, String, Route, List[(DataSet, String)]) => Unit) = {
     it("when searching for it directly") {
       def innerTermExtractor(dataSet: DataSet) = outerTermExtractor(dataSet)
-        .filterNot(term => Generators.luceneStopWords.exists(stopWord => term.trim.equalsIgnoreCase(stopWord)))
+        .filterNot(isAStopWord)
 
       doTest(innerTermExtractor)
     }
@@ -134,7 +136,7 @@ class LanguageAnalyzerSpec extends BaseSearchApiSpec {
           term.toLowerCase.endsWith("ies") ||
           term.toLowerCase.endsWith("es") ||
           term.toLowerCase.endsWith("y")) // This plays havoc with pluralization because when you add "s" to it, ES chops off the "es at the end
-        .filterNot(term => Generators.luceneStopWords.exists(stopWord => term.trim.equalsIgnoreCase(stopWord)))
+        .filterNot(isAStopWord)
         .flatMap {
           case term if term.last.toLower.equals('s') =>
             val depluralized = term.take(term.length - 1)
@@ -147,7 +149,7 @@ class LanguageAnalyzerSpec extends BaseSearchApiSpec {
               Some(pluralized)
             } else None
         }
-        .filterNot(term => Generators.luceneStopWords.exists(stopWord => term.trim.equalsIgnoreCase(stopWord)))
+        .filterNot(isAStopWord)
 
       doTest(innerTermExtractor)
     }
@@ -163,7 +165,7 @@ class LanguageAnalyzerSpec extends BaseSearchApiSpec {
             val terms = getIndividualTerms(innerTermExtractor(dataSet))
               .filter(_.length > 2)
               .filterNot(term => Seq("and", "or", "").contains(term.trim.toLowerCase))
-              .filterNot(term => Generators.luceneStopWords.exists(stopWord => term.trim.equalsIgnoreCase(stopWord)))
+              .filterNot(isAStopWord)
 
             if (!terms.isEmpty) {
               val termGen = for {